remove compat_stream.lua
